requirements.txt
python-telegram-bot~=21.11.1
PyYAML~=6.0.2